+++ virtualenvwrapper/hook_loader.py
@@ -125,7 +125,7 @@ def run_hooks(hook, options, args, output=None):
             continue
         plugin = ep.load()
         if options.listing:
-            print '  {0:10} -- {1}'.format(ep.name, inspect.getdoc(plugin) or '')
+            print '  %-10s -- %s' % (ep.name, inspect.getdoc(plugin) or '')
             continue
         if options.sourcing:
             # Show the shell commands so they can